requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dis==1.0.1
dissononce==0.34.3
PyJWT==2.7.0
idna==3.10
noiseprotocol==0.3.1
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.4
urllib3==2.4.0
pytest==8.2.2
pytest-cov==5.0.0 